Update patch set 1
Patch Set 1:

Shirayuki could you please explain the workflow for how and why you grep for keys. I would be really keen to understand this better. There is alread

It seems like there would be a better way then adding comments in javascript which are just confusing to a non-translator.

$ grep -r 'mobile-frontend-photo-nag-2-bullet-1-heading' .

Shows me the following:

./includes/Resources.php:			'mobile-frontend-photo-nag-2-bullet-1-heading',

If i then go to includes/Resources.php I see the message preceded with "// mf-photo.js" and can tell it belongs to mf-photo.js

Is this really too much effort?
Why is there a need for comments to be in the javascript as well?

Patch-set: 1