Update patch set 1
Patch Set 1:

I want the message documentation to be clarified about the usage ("This message is used as ...") eventually.

"// mf-photo.js" seems to be useful enough.

Could you describe about the usage in the message documentation (on translatewiki.net)?

e.g. "Used as label for the button", "Used as link text", "Used as tooltip for the link" or something.

Patch-set: 1