Update patch set 1
Patch Set 1:

So the problem is actually that you are having to look at the code in the first place?

If the qqq codes explain whether something is used as a label or attribute this prevents you having to grep in the first place? Ideally I would have thought that translators would like to be code agnostic - ie. never have to grep and look at code. Is this a fair assessment Shirayuki?

Patch-set: 1