Update patch set 1
Patch Set 1:

It will go on until your location can be found - not an eternity.

I'm not sure how we can improve this.

We could change the wording to "Attempt #3 to determine your location". This will however introduce a new requirement for translations.. which would suck and delay promotion to stable.

Patch-set: 1