Run all mobile QUnit tests
Remove filter=MobileFrontend default so that all things that run on
mobile (even if they don't belong to MobileFrontend itself) are tested.
This increases the number of tests from 198 to 217 as of this commit, so
there's no significant performance decrease.

Also, add an optional argument to scripts/qunit.sh, so that you can
filter the tests yourself when developing, e.g.

    scripts/qunit.sh nearby

will only run tests related to Nearby.

Change-Id: I3406aa2e950a7fd88818d7c847082249d287022b