Update patch set 1
Patch Set 1: Code-Review-1

May. I've asked for new asset. What change occurred to this image? It's not clear from the commit... we should let May know.

Let's hold off merging this till the last minute in case we get a new asset and need to change that as well..

Patch-set: 1
Label: Code-Review=-1