Update patch set 6
Patch Set 6:

I fixed the declaration of $device.

After looking over all of this again, I couldn't help but wonder... why don't we have phpunit tests for our skins? i dont know if this patchset is the right place to add them, but it would be really nice if we had them and wrote methods in our skins such that they were easily testable.

Patch-set: 6