Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

The Go back button at the top looks strange. It is only partly round. And it's the third kind of back button we have now.

Also, toggling does not seem to be necessary there and does not work as expected if you add a new section within a section.

Patch-set: 2
Label: Code-Review=-1