Update patch set 2
Patch Set 2: Verified+2 Code-Review+1

This looks OK, but should also update documentation for the varaible in MobileFrontend.php. We should remove the 'site' key, and also include documentation about the 'copyright' key that's currently in use in production. Also, the Also, is there a config change for production to reflect this?

Incidentally, does anyone remember the logic for using the 'site' key? I can't, and nothing is obvious in the code... but perhaps that's the impetus for getting rid of this :)

Patch-set: 2
Label: Verified=+2
Label: Code-Review=+1