Update patch set 1
Patch Set 1:

Thanks.

Will have to rebuild hook implementation in order to make this work. The migration attempt from last Friday and earlier this week half worked in that it was producing multiple copies of the banner even though it would seem that shouldn't have been the case; I guess you could say it doubly worked!

This is on our radar.

Patch-set: 1