Update patch set 1
Patch Set 1: Code-Review-1

Overall I like where you're headed with this. I don't think the commit message quite lines up with what's happening here.

Also, this makes it so the only use of $wgMobileFrontendLogo is in includes/skins/UserLoginAndCreateTemplate.php - which is kinda silly.

What if we changed $wgMFCustomLogos to $wgMFLogos, and in that defaulted to some generic MW logo (like the one you delete in this patchset?)

Patch-set: 1