Update patch set 1
Patch Set 1: Code-Review-1

Something not quite right with this. On iOS the margin is now increased between the image and the first heading (I spent much time with Vibha perfecting that)

It also has the effect of making the text overlap the progress bar at the bottom with 3 circles.

Patch-set: 1
Label: Code-Review=-1