Update patch set 1
Patch Set 1:

Jon convinced me that his approach is better - mostly because if we rely on the factory method, then all possible Skin classes would *also* need to have a factory method, which is silly. My only recommendation here would be to invert the if() logic - check for the WML content type (rather than HTML), so functionally we default to HTML rather than WML (since WML is the unique case).

Patch-set: 1