Update patch set 2
Patch Set 2:

I inverted the if() logic and removed the class_exists check. If the class doesn't exist, the whole thing should break anyway. Alternatively, we could fall back to desktop site if the class doesn't exist, but we'd probably want to generally clean up the logic in the function if we prefer to take this route.

Patch-set: 2
Reviewer: Awjrichards <91@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0