Update patch set 3
Patch Set 3:

We should probably also use the following:
https://developer.mozilla.org/en-US/docs/Web/API/window.onbeforeunload
http://jonathonhill.net/2011-03-04/catching-the-javascript-beforeunload-event-the-cross-browser-way/

Back behaviour is a headache to get right...
So there's nothing wrong with the code but there are lot more situations we probably need to think more about:

jdlrobson: if i click the edit button to load the editor and then refresh the page or share the link - what is the expected behavior?
[10:47am] jdlrobson: *behaviour
[10:48am] Maryana: share the link should probably not go to the edit view, but i doubt it's going to be a very common scenario
[10:48am] Maryana: refresh the page - hmm. probably stay in the editor?
[10:51am] Maryana: i can definitely see the latter happening. some person who's never edited a wiki before will open it, be like "what are all these curly brackets? is it broken?" and try to refresh
[10:52am] jdlrobson: so currently both happen
[10:52am] jdlrobson: (but the edit is lost)
[10:52am] Maryana: eep
[10:52am] jdlrobson: and the only way to escape is to close
[10:52am] Maryana: we should warn the user that they're abandoning their changes, then
[10:52am] jdlrobson: actually i lie..
[10:52am] jdlrobson: if you go to http://localhost.beta/w/index.php/Duel_Masters#editor-0 and click back you go to the page you were on before - so there is no way of seeing the article
[10:53am] jdlrobson: isn't back behavior fun 
[10:53am] Maryana: double eep
[10:53am] jdlrobson: (likewise with hitting cancel)
[10:53am] Maryana: triple eep
[10:53am] jdlrobson: you can still edit though and save
[10:53am] Maryana: so to see the article you have to make an edit? heh. well, that's one way to make people contribute!
[10:54am] jdlrobson: but it doesn't take you back to the article 
[10:54am] Maryana: aaahh
[10:54am] jdlrobson: funny enough it takes you back to the editor
[10:54am] Maryana: haha
[10:54am] jdlrobson: with the message your edit was saved
[10:54am] Maryana: ok, that's weird. let's not do that.
[10:54am] Maryana: infinite edit loop!

It seems to me that what we want is the behaviour of the hash to be only linked to the current user and meaningless as soon as you hit refresh or share the link. Breaks my brain.

So the question is should we merge this and try it out, do an intense round of design and define this better and then either rewrite it / change it OR go back to the drawing board do this design work again and come back to it?

Maryana: i think we should merge what we've got and then do an intense test round with design
[11:03am] Maryana: trying out all possible scenarios
[11:03am] Maryana: back button party, if you will 
[11:05am] Maryana: it could be that it works 90% of the time, and 10% of the time it's weird… and maybe we can live with that
[11:05am] Maryana: or it could be that it works 60% of the time and the other 40% is weird and needs another mental model or something

Juliusz what do you think?

Patch-set: 3