Update patch set 1
Patch Set 1:

Are you sure about the upload count thing?
I thought the purpose was to hide it so it never says 0 and reveal it when you upload the first image so that the user sees an upload count of 1 afterwards. Of are you saying the javascript conditionally inserts the count?

/me to lazy to check ;-)

Patch-set: 1