Update patch set 1
Patch Set 1:

This looks promising a few discussion points I'd like to make

=== Where code belongs ==
This code should not be in MinervaTemplate (templates should contain no logic - they should just render)

I can see two potentially better places for this
1) SkinMinerva
2) Our own extended version of OutputPage / ParserOutput
When placing things in our code base we should think about would I upstream this to core at some point to the thing it extends? If the answer is no then we shouldn't be putting code in that place and we should be putting it elsewhere. To me these transformations feel like they should belong to ParserOutput or OutputPage, so I wonder if we should do the 2nd.
(this diagram describes how I currently see the MobileFrontend -> MediaWiki core migration) http://imgur.com/2ZohDTa

== Output format ==
Instead of outputting HTML we should make SkinMinerva create a template variable similar to:
https://gist.github.com/jdlrobson/5439509#file-gistfile1-diff-L52

Something that resembles this:
array( 'sections'=> array( '1'=>array( 'id'=>1, text='hello' ), 'lead'=>'' )

MinervaTemplate then simply has to render it without adding any logic. This makes this code the most flexible from a frontend perspective by not making any assumptions about what we want to do with the output (the back to top link for example can be easily generated or not generated).

Patch-set: 1