Update patch set 2
Patch Set 2:

Note that this failure has been in this commit from the beginning. It wasn't visible throughout because "recheck" triggers the check pipeline (linting +1) not the test pipeline (lint+test +2). To trigger the test pipeline, right now you'll have to rebase it or edit the commit message (anything that creates a new patch set from a trusted author).

Patch-set: 2