Fix position-fixed workaround for older devices in overlays
Could be observed in the editor. The bottom bar would stay in the middle
of the page.

It also fixes the overlays on iPhone 3G (the overlay was not covering
the web site below it) and almost fixes the position: fixed fallback on
this phone (it didn't work, now it does, although the bottom bar does not
stick to the bottom, but almost to the bottom).

Change-Id: I957ddea76d64d406602a4bd80d64316a43bfc6bc