Update patch set 1
Patch Set 1: Code-Review-1

Thanks for the patch. This would be better done in JavaScript though and should be put in the alpha mode of the site first (all code must prove itself via data in alpha before being pushed to stable)

The toggle JavaScript module would be a good place to look.

Patch-set: 1
Label: Code-Review=-1