Update patch set 1
Patch Set 1: Code-Review-1

Thanks so much for the patch! I'm going to make a few minor changes to this to for you to review.

The main problem is we should make this experimental mode (alpha) only at first, so that we can experiment with the correct ux for this a bit more (we never push code straight to production)

Also since the language section is turned into a blue button categories should probably go above it otherwise the page looks strange.

Patch-set: 1
Label: Code-Review=-1