Make tapping more responsive
Until now we used the click event to make links and buttons do
something. This, unfortunately, causes a 300ms delays on all mobile
browsers between the actual tap (touchend) and the callback being
invoked.

This patch introduces a small library that adds a tap event for a more
responsive UI. On browsers without touch events it falls back to click
events.

I added it in alpha in several places for now: hamburger, overlay close
button, search input and page actions. If it works well, we could use it
everywhere instead of using click events.

Change-Id: Ib487355d86803fb09f169b573575b2a34f456ada