Change settings page heading
"Mobile site settings" is way too long and verbose. It leads to
horrendous translations in other languages (spanning across 2 or 3
lines). Settings is the name of the option in the menu and the heading
should be consistent. Everyone browsing the mobile site knows that when
they open settings, it's mobile site settings.

Change-Id: If43725f4311968c778309220d4581cd8a14f7fc9