Update patch set 2
Patch Set 2: Code-Review-1

Although this allows me to edit in the test case mentioned in the bug, the preview is messed up (the order of headings is changed). Is this OK?

I'm not even sure why we need a workaround for that. What is going on there? H1 is treated as a beginning of a section? Maybe we should rethink what a section is instead then?

Patch-set: 2
Reviewer: JGonera <542@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1