Update patch set 3
Patch Set 3: Code-Review+1

Although the design treatment of this is not perfect it does make it obvious that you are looking at an old revision.

It also will only show the subtitle when oldid is set in query string. This allows us to think of things like redirects separately.

Patch-set: 3
Label: Code-Review=+1