Update patch set 1
Patch Set 1:

(1 comment)

I'm not sure how this breaks animations - can you elaborate? Surely it's a case of simply re-rendering it and then showing it. This hints that there is some bad code somewhere upstream.

Refreshing views is likely to be something we are going to want to so I'd encourage us to spend more time on this and making a more generic way of doing this.

Was going to talk to you on irc about this but you signed off. Let's talk about this some more.

Patch-set: 1