Update patch set 4
Patch Set 4:

This doesn't seem to work in my local install.

In Safari in iOS simulator and in Chrome, I just get a black screen when clicking on an image, with an 'X' close box in the upper-right corner.

Both tested in 'experimental' mode.

Looks like it's making assumptions about the locations of files:

http://hawkeye.local/core/images/e/ef/San_Francisco_Museum_of_Modern_Art_in_2011.jpg 

In this case the actual file lives on Commons:
http://upload.wikimedia.org/wikipedia/commons/e/ef/San_Francisco_Museum_of_Modern_Art_in_2011.jpg

Sounds like something's trying to construct a URL manually instead of hitting the API to get the URL...

Also, loading the full-resolution image is a really bad idea -- many photos will be multi-megabyte multi-megapixel JPEGs, and some will be tens of megabytes (large TIFF or PNG images of scans, for instance). Others might be formats not supported natively by the browser, which require rendering as PNG or JPEG. JPEGs with EXIF rotation may also display incorrectly in a browser, while thumbnails provided by MediaWiki are pre-rotated as necessary.

Instead, I recommend asking the API for a thumbnail in the requisite screen size (times screen density), and no larger.

Patch-set: 4