Update patch set 6
Patch Set 6:

Now works as expected with Commons images -- looking good!

Much nicer on Retina display as well (tested in iOS simulator).

Close icon still looks pixely but we'll leave that to design. :)

I do note that transparent images don't always look good on the black background -- this could be a problem for many diagrams and charts. Making just the <img> have a white background may resolve this by giving transparent images a light BG over the actual image area, without removing the black borders *around* an opaque image.

Zooming seems hit-or-miss; looks like it's just allowing the viewport to be zoomed depending on the browser.

I notice that tapping on the image seems to pop up the caption & 'details' button at the bottom. I would expect tapping in the center of the image a second time to hide it again.

Patch-set: 6
Label: Code-Review=0