Hygiene: Remove obsolete code from mf-toggle-dynamic.js
checkHash() was redundant. Dynamic section loading is already handled in
section-toggle event callback (mf-toggle-dynamic.js) and scrolling to the
section is handled in wm_reveal_for_hash() (mf-toggle.js).

Previously we would serve the HTML for the lead section and section headings.
If a user navigated to Page#Foo and #Foo was hidden in say section 2,
section 2 would get lazy loaded and suddenly exist,
so then you'd have to honour the hash and scroll to it.
Since we don't do that anymore
(the entire page is served by default)
code no longer needed!

Change-Id: I47a51a765503e15ab0efe31d63f4e991bc2df579