Update patch set 1
Patch Set 1: Code-Review+2

So I remember what this was attempting to do now... previously we would serve the HTML for the lead section and section headings. If a user navigated to Page#Foo and #Foo was hidden in say section 2, section 2 would get lazy loaded and suddenly exist, so then you'd have to honour the hash and scroll to it.

Since we don't do that anymore (the entire page is served by default) I think this is not an issue but may be if we explore reducing the initial page load again.

Patch-set: 1
Label: Code-Review=+2