Using our own copyright i18n message for mobile
The copyright i18n message in core is commonly overriden on-wiki
with paragraph-long licensing information. Having a licensing
message that long is not suitable for mobile display. This allows
us to have a separate version for mobile views.

Also making it so that if no licensing info is present in the
config, it doesn't display a broken message.

Change-Id: Id082230532b3857cfb015939c98a62eda4c3c59f