Update patch set 2
Abandoned

Any patchsets open for 2 months with no activity in last 2 weeks will be closed.

Since this hasn't had any recent activity we are going to close it to keep the backlog fresh. Please do not take this as a rejection of your idea or work, simply as a way of keeping our process as clean as possible.

Please feel free to fix up and resubmit this patch - it will still exist on Gerrit!

Patch-set: 2
Status: abandoned