Update patch set 3
Patch Set 3:

I realize we don't have phpunit tests for the ApiMobileView class yet, but it would be nice if we did. It would be particularly nice to have a test for this particular change as it seems like there's potential for future breakage.

Patch-set: 3
Reviewer: Awjrichards <91@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0