Update patch set 4
Patch Set 4: Code-Review+2

(5 comments)

Various moans but I'm happy to +2 in the good faith as long as these get fixed in a follow up request as I think the thing this achieves is much much greater than my nags.

My main concern is the addition to the common.less file - this should not be in there unless it is a common pattern across the site which currently it isn't. Please move into editor.less or apply the pattern somewhere else.
w

Patch-set: 4
Label: Code-Review=+2