Migrating Mobile thanks into Thanks extension
Adding mobile.stable.common as a dependency for
mobile.mobilediff.scripts since we need it for the toast and
mediawiki.jqueryMsg. Plus we'll need more of it in the future
for support of profile features.

Also changing the Thank button to green per Vibha. Hardcoding it
for now, but eventually will use mediawiki.ui instead.

See also: change Ifaf44fe8

Change-Id: I56e2c5bc69f85e83ab3dfd9b9e617dbb98661870