Update patch set 1
Patch Set 1:

I talked to Arthur who added this call and he doesn't remeber why it's there, so it is safe to remove it without adding an exception handler - I'd like a SQL exception to propagate far enough to be logged instead of being silently ignored. Will tweak this in a second.

Patch-set: 1
Label: Verified=0