Update patch set 7
Patch Set 7:

Templates do not depend on MobileFrontend JavaScript at all. If left in this form other extensions are able to make use of them simply by checking for the existence of the module or whether MobileFrontend is installed and adding modules which list mediawiki.template as a dependency without the penalty of having to include other JavaScript.

Thus separating them in this fashion is a good thing from my point of view.

Patch-set: 7
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0