Update patch set 2
Patch Set 2: Code-Review-1

So Kaldari and I spoke about this.
So my interpretation  of the discussion was this:

We discussed workflows and how caching watchlist and diff css as same time would be useful as a kind of prefetching.

Profile and uploads css would be useful to group at the same time as profile pages link to uploads pages.

nearby could do with some refactoring

Not clear what way forward is, so happy to park this in favour of getting the other patchset that follows this merged.

I think there are advantages to minimising the http requests for stylesheets but certain pages /are/ very different to others so it might be useful to keep those stylesheets separate (e.g. mobilemenu/login)

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1