Update patch set 5
Patch Set 5: Code-Review-1

jdlrobson: MaxSem: I still worry this is error prone - we render the message 'mobile-frontend-profile-contributions' which assumes on the value 500 but the value 500 is set by MobileUserInfo.php - this seems bad no?
[2:51pm] jdlrobson: what if I update MobileUserInfo to cap to 600 how I know to update the message?

My suggestion would be to change the method name / parameters to make this more explicit:
either $this->userInfo->countRecentUploads( $fromDate, 500 );
$this->userInfo->countCappedRecentUploads

If either Kaldari, JGonera or Arthur disagree with me feel free to overrule my -1 - but I don't like it.

Patch-set: 5
Label: Code-Review=-1