Story 1269: AbuseFilter support
To test, configure the AbuseFilter extension and create the following
rule on Special:AbuseFilter:

Conditions: user_name = "<your test user name>"

Then check "Trigger these actions after giving the user a warning" to
test the warning behaviour or "Prevent the user from performing the
action in question" to check the disallow behaviour. Try editing with
the user name specified in the condition.

Bug: 52049
Change-Id: I557304105ba35ae8e760de5fe1adfbcbc9b5497e