Beta: Add Special:History
Provides Recent Changes and History functionality.
A basis to work off of.
Based loosely on the design in Story 184

Adding to beta as it is currently not linked to from
anywhere (this has been okayed by Kenan) - we cannot link
to it until it has paging

TODO: Design input needed on how to style deleted entries in history
(see FIXME notes in SpecialHistory)
Change-Id: I97f5aa3fb8b7a927defbababd9acc339c2611dbc