Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

Seems to be 2 things going on here - hygiene for message functions and the fix itself. Can these not be broken so the diffs are easier to grep?

Also we currently 404 when you give an invalid revision id on the mobile diff. I'd prefer we got into the habit of 404ing when a bad URL is given. Pages that 200 can be indexed by search engines so I'd prefer to do this..

Patch-set: 4
Label: Code-Review=-1