Update patch set 1
Patch Set 1: Code-Review-1

A -2 is harsh in my opinion. I'm sure this can be improved and is trying to fix a real code hygiene problem. I'd suggest you see the goals and provide an improved solution that allows inheritance of the Special:History page and avoids the code duplication we have everywhere. The diff increase is mostly due to the comments I have added on each of the functions so is not really a consideration here.

jdlrobson: MaxSem: -2 seems harsh - especially when i've added comments so of course the size is going to go up 
[2:14pm] kenanwang left the chat room. (Quit: kenanwang)
[2:14pm] MaxSem: jdlrobson, if (foo) { long block of text } else { long block of text } is bad
[2:15pm] MaxSem: an elegant solution would be to use inheritance
[2:15pm] MaxSem: though it's not very suited in this case
[2:15pm] jdlrobson: i still think it's an improvement - and if you see the follow up patch its of interest to be able to override these functions
[2:16pm] jdlrobson: another way would be to use private variables that get set elsewhere
[2:17pm] jdlrobson: my main annoyance was the logical for wl_namespace and rc_namespace - identical but in two different places
[2:17pm] jdlrobson: if you want i can remove all the comments to show the improved diff size

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1