Update patch set 2
Patch Set 2:

What's the rationale for this? Static functions are generally to be avoided, unless  there's a clear use case for this. For example, in this case you're not passing the context to messages, which might have repercussions.

Patch-set: 2
Reviewer: MaxSem <43@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0