Make View's defaults inheritable and extendable
Now when children of View define the `defaults` property, properties of
`defaults` of the parent are preserved if the child doesn't override
them.

Previous behaviour led to missing close button message in most overlays,
which suggests that extending defaults in this new way may be more
intuitive, if slightly more magical.

Change-Id: Id6bb5a89dea33200ea9cf35535644bde54d8820e