Update patch set 3
Patch Set 3:

I assume this is for the benefit of desktop browsers? I have not much knowledge of ARIA support in mobile but with limited use of a keyboard on mobile devices will this have any benefit for those users?

Also FYI, there are experiments with changing the way we do toggling which will impact this work if successful: https://gerrit.wikimedia.org/r/#/c/92555/

Patch-set: 3