Story 1289: Promote image viewer overlay to beta
If CommonsMetadata extension is present and is able to fetch the license
information and return it in the API response, then we show the license.
As a fallback we show a "License information" link pointing to the file
page, just like desktop media viewer does.

Note: license information is only shown when you click the image.

Change-Id: Iab2b857d009f25f077edb940d836ad84d576a9b8
Dependency: Ibaf09360041408e9bf9692056572f5b23b7f7c08