Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

This has me super confused for a while as I looked at Overlay.js hide function and the hide function didn't take any parameters.

I eventually realised PhotoUploadProgress is the only function to have a parameter to the hide method. This is very confusing and should at least have a comment to explain what is going on or even better should be part of PhotoUploadProgress itself...

Patch-set: 1
Label: Code-Review=-1