Update patch set 3
Patch Set 3:

I moved the override into MobileFrontendHooks::onReqestContextCreateSkin() after we are sure we're in mobile mode. Tested with ULS installed and this seems to work - preserving the setting on desktop but overriding it in mobile, and resolving the bug.

Patch-set: 3
Reviewer: Awjrichards <91@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0