Update patch set 7
Patch Set 7: Code-Review-1

This seems to work... ish. It loads the tutorial beneath the fold. When I tap on 'contribute an image', i see a brief flash, then the standard Special:Upload view (with the chrome, and the upload button). But if I scroll down, I see the upload tutorial. Tested on iOS simulator with 4" iPhone retina

Patch-set: 7
Label: Code-Review=-1