Update patch set 1
Patch Set 1: Code-Review-2

this reverts work we are doing to provide a mobile friendly history view. We override the watchlist special page on mobile and this should be no different. The existing state of the code will only apply to mobile in beta - the hook ensures it only applies in mobile mode not desktop. I'll provide more through information when at a computer but this code should not be merged. Note the alias work is useful though.

Patch-set: 1
Label: Code-Review=-2