Update patch set 1
Patch Set 1:

(1 comment)

With the change listed this will be mergeable.
Apologies for the short reply yesterday I was on a mobile phone making explanation difficult.

In the current form it would break the license agreement in that there would be no working link in beta and alpha to page history.

Max just download it locally and see for yourself.

Patch-set: 1