Update patch set 2
Patch Set 2:

Define fishbowl wiki... maybe I've misunderstood it? I was assuming it was a wiki where all users have to an account in order to edit? If so this patch /does/ fix that problem. If not then please can you clarify what this means?

e.g. $wgGroupPermissions['*']['edit'] = true;

Patch-set: 2