Update patch set 4
Patch Set 4:

Most of the tests do not work against a freshly installed wiki. Either because of Watir timeout or because the test is expecting something which is only on some specific wiki.

The features definition need a bunch of improvements before we can trigger the tests from Gerrit.

Patch-set: 4
Label: Verified=0