Update patch set 7
Patch Set 7: Code-Review-1

This just doesn't look right... I don't think it's a change that can be made so lightly. From what I heard we will probably work on this in the next iteration. I'm not sure if there's a point in doing it in alpha just for two weeks.

Also, if you really insist, this breaks a lot of styling. Margins in the watchlist are wrong, headings have unnecessary letter-spacing (asked designers about that). I know alpha is for experiments, but let's not make it ugly.

Patch-set: 7
Label: Code-Review=-1