Update patch set 5
Patch Set 5:

Special:RecentChanges is a history feed for ALL pages. It therefore redirects to Special:History

Special:Contributions is a history feed that allows filtering by username
e.g. Special:Contributions/Jdlrobson shows all history of the user Jdlrobson
It doesn't redirect to Special:History as it relies on the page to know that everything after the slash is a username not a page.

Patch-set: 5
Label: Verified=0