Update patch set 3
Patch Set 3:

Chris, Jeff, please merge this into master if you do not have any complains.

If nobody complains in the next day or two, I will merge this into master. It touches a lot of code and it is likely to have a lot of conflicts if it is not merged soon.

Patch-set: 3