Story 1448: New KeepGoing workflow for adding links
Enable with $wgMFKeepGoing = true

This is a new workflow to encourage users to add links to articles.
After the user makes their first edit, it asks them if they want to
add links to a different article. If they agree to, it takes them to
a random article in the category All articles with too few wikilinks
and shows them instructions for creating links (only once). If they
edit that article as well, it asks them if they want to fix another
article, and the cycle repeats (without the instructions).

The functionality of the random page drawer (in Alpha) has been left
exactly how it was for now (although this does create some
duplicated code). Once we decide if we are keeping either or both of
these features, we can abstract the code as is appropriate.

Change-Id: Idd82e88302f3ecda0da261c48b2a19c555feaa8b