Update patch set 1
Patch Set 1: Code-Review-1

As stated in a previous patchset that got merged despite my -1 this is getting confusing. There is already a README in the main directory with even more detailed instructions on how to do this.
Please instead port that here and point to the README from it rather than reinventing the wheel!

Also the other README is written in mediawiki markup not markdown. This is a little weird but I guess we can live with it if this is preferred.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1