More robust permission checking for User Profile
The previous permission checking on the client-side didn't work for
fish-bowl wikis or blocked editors, and on the server side it
didn't work in certain rare situations (like the UserProfileIntro
being protected). We now make sure the user can actually edit the
page in both cases.

Also, since we're generating the UserDescriptionPage's title
earlier now, change getWikiPageText() to take advantage of that.

Change-Id: I2d59153ac41e30239641380c213c7d85db852d29