Moved information on how to run Selenium tests to a separate README file
The goal is to make it easy for people to start contributing to
Selenium tests.

While working on resolving a couple of bugs (see below) we are updating
existing documentation and adding README files to tests/browser folders
of all repositories that contain Selenium tests. Some people will land
there and never notice the main README file.

Bug: 57841
Bug: 56089
Change-Id: I814f7167ef5aafca62fc92d11ad27956cc3a4d89