Override language_urls template data to avoid hook abuse
Various extensions abuse the SkinTemplateOutputPageBeforeExec hook
to append things that are not language urls to this template variable.
Mobile expects these all to be actual languages - not links to edit languages or other 'phantom' list items.

We only recently started using core to generate this list 2 weeks ago
and since then we've hit various bugs so it's obviously not in a usable
state. As a result we should revert back to the tried and tested way of 
2 years of generating this list ourself after the hook has run to ensure we get what we need.

Bug: 58102
Bug: 57091
Bug: 57094
Change-Id: I09a870aa8e2a8b9be8c45b54d5c646ae5b105cf9