HYGIENE: Deprecating @baseFontSize
We are reinventing the wheel here. The unit 1em means the browser's
base font size (as set in the preferences). So 2em is just a
multiplier on that. There is no reason for us to duplicate this
logic by redeclaring our own baseFontSize that is also 1em and
multiplying everything by that. In essence we are reimplementing
the functionality of the em unit.

And if we do actually want to change all of our font sizes, we
would adjust @fontScalingFactor anyway, not @baseFontSize.

In converting our own multiplier to ems, I rounded to the nearest
tenth in some cases, which should have negligible impact. The only
exception to this is bodyLineHeight which I converted to a pure
multiplier (which is better to use for a default line-height due to
inheritance), and rounded up instead of down since 1.3 is a rather
cramped line-height and most of our text is set to much larger
line-height values anyway (1.5 or 1.65)

Change-Id: Ib9864ba6c951cf3a30bd45c8bbb5aa7d36ab8d4a