Update patch set 5
Patch Set 5:

Jeff, I don't think the comment saved correctly, but I did read it. In line 44 of the steps file I am explicitly checking that the href link string of the top image link in the page contains the random_string.  This tells me that the file was saved correctly and also that the link to the file is the top link on the page.

Patch-set: 5
Label: Verified=0