Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

Jon and I need to refactor this. handleFile() should probably be moved somewhere so that it can be used in more than one place and we should probably have two different routes (hashes), one for tutorial with a file input and another for tutorial with a close button.

Patch-set: 5
Label: Code-Review=-1