Update patch set 4
Patch Set 4: Code-Review+2

Although this is an evil hack and I share Parent5446's concern that it could cause unintended consequences, it seems like the best solution for now (until there is traction on bug 57091 and bug 57094). Since this fixes a live bug and does not introduce any concrete problems in the meantime, I'm going to go ahead and merge it so it can go out on Tuesday. In the meantime, we should continue discussing better solutions (either on wikitech-l or in the bugs).

Patch-set: 4
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2