Update patch set 3
Patch Set 3:

> Also, IMO, we should explicitly set $wgAllowUserJs and $wgAllowUserCss to false in MobileFrontend.php in the config variables section (so that it is more obvious that these options are available).

No, that's not a good idea. Those are MediaWiki configuration variables, and setting them unconditionally in an extension is bad. A better idea is when this patch is merged to document this feature on [[mw:Extension:MobileFrontend]] (or other page) so people are aware of it.

Patch-set: 3
Reviewer: Legoktm <458@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0