Story 1483: Make overlay headers fixed
Introduce .overlay-header-fixed-container to avoid a bug which caused
.overlay-header background not to cover the whole header in Chrome when
a combination of box-sizing: border-box, width: 100%, padding, and
display: table was used. It might also be useful in future to group and
additional element which should remain fixed together with the header.

Modify @overlayHeaderMargin to avoid weird rounding errors (in search
overlay button in the header and button in "search within pages" had
different width when it was 0.4em).

Modify the header message to allow truncating page title even if it's
not at the end.

Change-Id: I0c5bc2426c24ab15646b840c407713ee1228f5ab