Update patch set 4
Patch Set 4:

Correct, this still shouldn't work.  As of yesterday, I completely changed the way we handle events inside target modules.  I'll be adding the bits needed ( probably to this patch ) to bring it up to speed.

Change https://gerrit.wikimedia.org/r/#/c/100520 is in preparation of making this work, and once it does i will surely add it as a dependency.

Patch-set: 4
Label: Verified=0