Update patch set 1
Patch Set 1:

I support the idea, but I'm not sure a mobile-specific Special Page is really the right solution to the problem. The design team has thrown around the idea of a personal Wikipedia "home page" or "landing page", which would consist of configured dashboards rather than MediaWiki templates and could exist in both mobile and desktop. If such a page existed, I imagine we would want to direct mobile users to it by default rather than the Main Page.

In the meantime, however, I'm undecided as to what the best approach would be. Perhaps we should have an RfC ;)

Patch-set: 1
Label: Verified=0