Avoid using router for Issues overlay when we are in Editing overlay
This is a quick fix for bug 58304 until the Overlay Manager is
implemented. This prevents the Issues overlay from being initialized
with M.router when we are already in the Editor overlay. Otherwise,
closing the Issues overlay would intantiate a new Editing overlay
and the user would lose the edit they were working on.

This also fixes the bug of getting two warnings about abandoning
your edit.

Two related bugs will be fixed in follow-up changes:
* Preventing the number of issues from increasing each time the
overlay is opened
* Preventing the initial warning about abandoning your edit

Bug: 58304
Change-Id: I76fbe210a453c4fb10c92b125ed4676c8d0b769d