Docs: Document overlays with KSS
* Reverse order of stylesheets so kss.less can hijack styles to
make them work in the KSS documentation guide
* Make position-fixed class important and drop that from overlay-bottom
to make this clearer and allow position fixed handling to work
* Use @wgMFMobileImageRootDir to get round problem with image file
paths - the style guide css is generated in less/public and all image
paths without this will 404. This allows us to change the file path
of the overlay submit and continue and cancel buttons so that they
point to the correct place.

Change-Id: I05ce81dcab2d00f93cde3114452ea9dbeecce6f8