Add intermediate Special:MobileEditor to catch users without JavaScript
When following a link with action=edit, users will first be
redirected to Special:MobileEditor/Page?params=from&original=request.
If JavaScript is enabled in their browser, they will immediately
be redirected to Page#editor/<section> as intended. If JavaScript is
not available, a warning message will appear with a link to
return to the article in view mode.

Bug: 54475
Change-Id: I6c74a370c7850cea071696c38941da3946ac0e1c