Hygiene: Fixing default table styling
Since we are forcing tables to use display block instead of display
table, having a border on both the cells and the table looks weird.
We really only need borders on the th and td, which we already have.

Note due to the use of reset.css in MobileFrontend this will
revert to border none.

Change-Id: Icffcb9cb89e01b4450fd9c2ea197288c021b56ef